Moved deep-copy support to a trait.
